+++ R/supppr.R
@@ -1,0 +1,22 @@
+#' @title supppr
+#'
+#' @description supppr created from scratch
+#'
+#' @format A data frame with 240 rows and 8 variables:
+#' \describe{
+#'  \item{STUDYID}{Study Identifier}
+#'  \item{RDOMAIN}{Related Domain Abbreviation}
+#'  \item{USUBJID}{Unique Subject Identifier}
+#'  \item{IDVAR}{Identifying Variable}
+#'  \item{IDVARVAL}{Identifying Variable Value}
+#'  \item{QNAM}{Qualifier Variable Name}
+#'  \item{QLABEL}{Qualifier Variable Label}
+#'  \item{QVAL}{Qualifier Value}
+#' }
+#' @seealso \code{\link{ae}} \code{\link{ds}} \code{\link{dv}} \code{\link{ho}} \code{\link{ie}} \code{\link{mh}} \code{\link{pr}} \code{\link{suppho}} \code{\link{supppr}} \code{\link{ts}}
+#' @keywords datasets supppr
+#' @name supppr
+#' @examples
+#'  head(data("supppr"))
+"supppr"
+